requirements.txt
streamlit>=1.22.0
python-dotenv>=1.0.0